+++ jest.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  moduleFileExtensions: ['js', 'ts'],
-  testEnvironment: 'jsdom',
-  transform: {
-    '^.+\\.(scss|less|svg|png)$': './jest/style-mock.ts',
-    '^.+\\.[t|j]sx?$': 'babel-jest',
-  },
-};