+++ vite.config.ts
@@ -1,9 +1,0 @@
-import { defineConfig } from 'vite';
-
-export default defineConfig({
-  base: '/file-upload-with-preview/', // For GitHub docs support
-  build: {
-    outDir: '../docs',
-  },
-  root: 'example',
-});